refactor auth views into seperate file
