+++ backend/samfundet/view/general_views.py
@@ -9,7 +9,7 @@ from itertools import chain
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.request import Request
-from rest_framework.generics import ListAPIView
+from rest_framework.generics import ListAPIView, CreateAPIView
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet
 from rest_framework.decorators import action
@@ -39,6 +39,7 @@ from samfundet.serializers import (
     ClosedPeriodSerializer,
     OrganizationSerializer,
     SaksdokumentSerializer,
+    UserFeedbackSerializer,
     UserGangRoleSerializer,
     InformationPageSerializer,
     UserGangSectionRoleSerializer,
@@ -58,6 +59,7 @@ from samfundet.models.general import (
     Organization,
     Saksdokument,
     InformationPage,
+    UserFeedbackModel,
 )
 
 
@@ -216,3 +218,26 @@ class MerchView(ModelViewSet):
     permission_classes = (RoleProtectedOrAnonReadOnlyObjectPermissions,)
     serializer_class = MerchSerializer
     queryset = Merch.objects.all()
+
+
+class UserFeedbackView(CreateAPIView):
+    permission_classes = [AllowAny]
+    model = UserFeedbackModel
+    serializer_class = UserFeedbackSerializer
+
+    def create(self, request: Request) -> Response:
+        data = request.data
+
+        serializer = self.get_serializer(data=data)
+        serializer.is_valid(raise_exception=True)
+
+        UserFeedbackModel.objects.create(
+            user=request.user if request.user.is_authenticated else None,
+            text=data.get('text'),
+            path=data.get('path'),
+            user_agent=request.META.get('HTTP_USER_AGENT'),
+            screen_resolution=data.get('screen_resolution'),
+            contact_email=data.get('contact_email'),
+        )
+
+        return Response(status=status.HTTP_201_CREATED, data={'message': 'Feedback submitted successfully!'})
